+++ packages/create-plugin/__e2e__/e2e.test.ts
@@ -1,5 +1,6 @@
 import assert from "assert";
-import { executeCommandWithInteractiveInput } from "./utils/helper";
+import type { QuestionInput } from "./utils/executeCommand";
+import { executeCommandWithInteractiveInput } from "./utils/executeCommand";
 import {
   CREATE_PLUGIN_COMMAND,
   DEFAULT_ANSWER,
@@ -13,6 +14,7 @@ import {
   assertObjectIncludes,
   readPluginManifestJson,
 } from "./utils/verification";
+import { getBoundMessage } from "../src/messages";
 
 describe("create-plugin", function () {
   let workingDir: string;
@@ -22,21 +24,44 @@ describe("create-plugin", function () {
   });
 
   it("#JsSdkTest-1 Should able to create a plugin with specified output directory and required options successfully", async () => {
+    const m = getBoundMessage("en");
     const outputDir = "test1";
-    const ANSWERS = [
-      "test1-name",
-      "test1-description",
-      DEFAULT_ANSWER,
-      DEFAULT_ANSWER,
-      DEFAULT_ANSWER,
-      ANSWER_NO,
-      ANSWER_NO,
+    const questionInput: QuestionInput[] = [
+      {
+        question: m("Q_NameEn"),
+        answer: "test1-name",
+      },
+      {
+        question: m("Q_DescriptionEn"),
+        answer: "test1-description",
+      },
+      {
+        question: m("Q_SupportJa"),
+        answer: DEFAULT_ANSWER,
+      },
+      {
+        question: m("Q_SupportZh"),
+        answer: DEFAULT_ANSWER,
+      },
+      {
+        question: m("Q_websiteUrlEn"),
+        answer: DEFAULT_ANSWER,
+      },
+      {
+        question: m("Q_MobileSupport"),
+        answer: ANSWER_NO,
+      },
+      {
+        question: m("Q_enablePluginUploader"),
+        answer: ANSWER_NO,
+      },
     ];
+
     const response = await executeCommandWithInteractiveInput(
       CREATE_PLUGIN_COMMAND,
       workingDir,
       outputDir,
-      ANSWERS,
+      questionInput,
     );
 
     assert(response.status === 0, "Failed to create plugin");
