+++ packages/create-plugin/__e2e__/utils/constants.ts
@@ -3,13 +3,4 @@ export const ANSWER_YES = "Yes";
 export const ANSWER_NO = "No";
 
 export const CREATE_PLUGIN_COMMAND = "create-plugin";
-
-export const QUESTION_LIST = [
-  "Input your plug-in name in English [1-64chars]",
-  "Input your plug-in description in English [1-200chars]",
-  "Does your plug-in support Japanese?",
-  "Does your plug-in support Chinese?",
-  "Input your home page url for English (Optional)",
-  "Does your plug-in support mobile views?",
-  "Would you like to use @kintone/plugin-uploader?",
-];
+export const CREATE_KINTONE_PLUGIN_COMMAND = "create-kintone-plugin";
